# [24.2] Fix private role name performance issue