--- a/lib/galaxy/files/__init__.py
+++ b/lib/galaxy/files/__init__.py
@@ -396,7 +396,11 @@ class ProvidesFileSourcesUserContext(FileSourcesUserContext, FileSourceDictifiab
     def role_names(self) -> Set[str]:
         """The set of role names of this user."""
         user = self.trans.user
-        return {ura.role.name for ura in user.roles} if user else set()
+        role_names = set()
+        if user:
+            role_names = {ura.role.name for ura in user.roles}
+            role_names.add(user.email)  # User's private role may have a generic name, so add user's email explicitly.
+        return role_names
 
     @property
     def group_names(self) -> Set[str]: