--- a/lib/galaxy/model/db/role.py
+++ b/lib/galaxy/model/db/role.py
@@ -41,3 +41,12 @@ def get_private_user_role(user, session):
 def get_roles_by_ids(session: galaxy_scoped_session, role_ids):
     stmt = select(Role).where(Role.id.in_(role_ids))
     return session.scalars(stmt).all()
+
+
+def get_displayable_roles(session, trans_user, user_is_admin, security_agent):
+    roles = []
+    stmt = select(Role).where(Role.deleted == false())
+    for role in session.scalars(stmt):
+        if user_is_admin or security_agent.ok_to_display(trans_user, role):
+            roles.append(role)
+    return roles