--- a/lib/galaxy/model/db/role.py
+++ b/lib/galaxy/model/db/role.py
@@ -1,3 +1,5 @@
+from typing import Dict
+
 from sqlalchemy import (
     and_,
     false,
@@ -6,6 +8,7 @@ from sqlalchemy import (
 
 from galaxy.model import (
     Role,
+    User,
     UserRoleAssociation,
 )
 from galaxy.model.scoped_session import galaxy_scoped_session
@@ -50,3 +53,10 @@ def get_displayable_roles(session, trans_user, user_is_admin, security_agent):
         if user_is_admin or security_agent.ok_to_display(trans_user, role):
             roles.append(role)
     return roles
+
+
+def get_private_role_user_emails_dict(session) -> Dict[int, str]:
+    """Return a mapping of private role ids to user emails."""
+    stmt = select(UserRoleAssociation.role_id, User.email).join(Role).join(User).where(Role.type == Role.types.PRIVATE)
+    roleid_email_tuples = session.execute(stmt).all()
+    return dict(roleid_email_tuples)