Fixes #19654 

The gist of the issue: https://github.com/galaxyproject/galaxy/issues/19654#issuecomment-2672981287

TODO:
- [x] Add fix for notifications (select roles)

### Solution:

Intercept Role data after it has been retrieved from the database but before it's been sent with the Response. Make one call to the database retrieving a mapping of private role IDs to associated user emails, load into a set and use to augment the Role data as needed, replacing names of private roles with corresponding emails, or adding them as an additional field.

Note: there's much duplication here, but to get rid of it we'd have to touch many places (fastapi controllers, services, legacy controllers, managers), and that kind of refactoring should be done on the dev branch. I've tried to change only what's necessary to fix the bug.

### Discarded approaches:

- Set the private role's description field to "Private role for [user' current email]" at runtime. (That field is null by default in newer versions and "private role for [user email]" in older versions.) The value is not persisted and is derived from the associated user's current email.
Reason for discarding:  breaks logic on the client that assumes a private role name to contain a user's email.
(also applies for using any new non-mapped field, like "displayed_name')
- A relationship mapping added to the Role model definition that eagerly loads associated user's email. 
Reasons  for discarding: https://github.com/galaxyproject/galaxy/pull/19679#issuecomment-2688364283
- Use limit + pagination.
Reason  for discarding: requires major changes on the client (role names displayed in dropdowns, etc.)

### Performance
I've tested this locally on 40K users+roles+associations. The roles api endpoint on the release branch is indeed unusable (as reported in the linked issue), whereas this branch is fine.



## How to test the changes?
(Select all options that apply)
- [x] I've included appropriate [automated tests](https://docs.galaxyproject.org/en/latest/dev/writing_tests.html).
- [x] This is a refactoring of components with existing test coverage.
- [x] Instructions for manual testing are as follows:
 - Try to edit permissions for libraries, library datasets, datasets, histories; and in the admin: role permissions, group permissions (on create and edit): observe that role names are displayed as names for non-private roles and emails of associated users for private roles - both in the drop downs and the selected values. 
 

## License
- [x] I agree to license these and all my past contributions to the core galaxy codebase under the [MIT license](https://opensource.org/licenses/MIT).