@jdavcs 2025-02-21
This won't work. This breaks logic that assumes a private role name to contain a user's email.